requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy












